Moved Binance related Json classes from Models/Json to ../Json
