+++ Cryptodd/Binance/Json/DepthUpdateMessageJsonConverter.cs
