+++ Cryptodd/Binance/Orderbooks/Websockets/BaseBinanceOrderbookWebsocket.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks.Dataflow;
+using Cryptodd.Binance.Json;
 using Cryptodd.Binance.Models;
 using Cryptodd.Binance.Models.Json;
 using Cryptodd.Ftx.Models.Json;
